packages/adk_impl/pyproject.toml
[project]
name = "adk-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Eder Ignatowicz", email = "ignatowicz@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "common",
    "google-adk>=1.21.0",
    "litellm>=1.75.5",
]

[project.scripts]
adk-impl = "adk_impl:main"
adk-prompt-chaining = "adk_impl.patterns.prompt_chaining.run:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
common = { workspace = true }
